+++ src/main.cpp
@@ -142,6 +142,7 @@ int main(int argc, char* argv[]) {
         compute.init_mask();
         compute.init_color();
         compute.init_compute_normals();
+        compute.init_multires_diff();
         compute.init_remesh_select();
         compute.init_remesh_grow_selection();
         compute.init_remesh_mirror_selection();